Don't use reprepro for building the apt repo
Reprepro does not allow to have more than one version of a package in
the same repo.  Since we want to keep multiple version of the CLI
packages available, we can no longer use reprepro.

Another requirement that we have is that our repo is on a GCP bucket
(at least for the moment).  Therefore, all existing versions of the CLI
packages are on the bucket and not on the machine where we build the
repo.  We don't want to have to download every package from the bucket
to be able to build a new repo as this would take more and more time
as we accumulate versions.  Instead we want to build the repo
incrementally by simply adding a new package version to it.

After investigating different alternatives, mostly `aptly`, it became
apparent that building a repository incrementally was not easy to do
with existing tools.  On the other hand, building the repo manually
ourselves turned out to be relatively easy.

Therefore, this commit replaces `reprepro` with a homemade script
that allows us to add a new version to an existing repo without having
to have the entire set of packages on the local machine.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>